+++ app/upgrades/v19/upgrades.go
@@ -11,17 +11,14 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/module"
 	authkeeper "github.com/cosmos/cosmos-sdk/x/auth/keeper"
-	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 	bankkeeper "github.com/cosmos/cosmos-sdk/x/bank/keeper"
 
 	upgradetypes "github.com/cosmos/cosmos-sdk/x/upgrade/types"
 	"github.com/ethereum/go-ethereum/common"
-	evmostypes "github.com/evmos/evmos/v19/types"
 	"github.com/evmos/evmos/v19/utils"
 	erc20keeper "github.com/evmos/evmos/v19/x/erc20/keeper"
 	erc20types "github.com/evmos/evmos/v19/x/erc20/types"
 	evmkeeper "github.com/evmos/evmos/v19/x/evm/keeper"
-	evmtypes "github.com/evmos/evmos/v19/x/evm/types"
 	stakingkeeper "github.com/evmos/evmos/v19/x/staking/keeper"
 )
 
@@ -50,8 +47,6 @@ func CreateUpgradeHandler(
 		if err := RemoveOutpostsFromEvmParams(ctxCache, ek); err == nil {
 			writeFn()
 		}
-
-		MigrateEthAccountsToBaseAccounts(ctx, ak, ek)
 
 		// run module migrations first.
 		// so we wont override erc20 params when running strv2 migration,
@@ -180,27 +175,4 @@ func registerERC20Extensions(ctx sdk.Context,
 	})
 
 	return err
-}
-
-// MigrateEthAccountsToBaseAccounts is used to store the code hash of the associated
-// smart contracts in the dedicated store in the EVM module and convert the former
-// EthAccounts to standard Cosmos SDK accounts.
-func MigrateEthAccountsToBaseAccounts(ctx sdk.Context, ak authkeeper.AccountKeeper, ek *evmkeeper.Keeper) {
-	ak.IterateAccounts(ctx, func(account authtypes.AccountI) (stop bool) {
-		ethAcc, ok := account.(*evmostypes.EthAccount)
-		if !ok {
-			return false
-		}
-
-		// NOTE: we only need to add store entries for smart contracts
-		codeHashBytes := common.HexToHash(ethAcc.CodeHash).Bytes()
-		if !evmtypes.IsEmptyCodeHash(codeHashBytes) {
-			ek.SetCodeHash(ctx, ethAcc.EthAddress().Bytes(), codeHashBytes)
-		}
-
-		// Set the base account in the account keeper instead of the EthAccount
-		ak.SetAccount(ctx, ethAcc.BaseAccount)
-
-		return false
-	})
 }