+++ testutil/integration/evmos/utils/contracts.go
@@ -9,6 +9,8 @@ import (
 	"github.com/evmos/evmos/v19/testutil/integration/evmos/factory"
 
 	abcitypes "github.com/cometbft/cometbft/abci/types"
+	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
+	evmostypes "github.com/evmos/evmos/v19/types"
 	evmtypes "github.com/evmos/evmos/v19/x/evm/types"
 )
 
@@ -30,6 +32,19 @@ func CheckTxTopics(res abcitypes.ResponseDeliverTx, expectedTopics []string) err
 		if !slices.Contains(availableLogs, expectedTopic) {
 			return fmt.Errorf("expected topic %s not found in tx response", expectedTopic)
 		}
+	}
+	return nil
+}
+
+// IsContractAccount checks if the given account is a contract account
+func IsContractAccount(acc authtypes.AccountI) error {
+	contractETHAccount, ok := acc.(evmostypes.EthAccountI)
+	if !ok {
+		return fmt.Errorf("account is not an eth account")
+	}
+
+	if contractETHAccount.Type() != evmostypes.AccountTypeContract {
+		return fmt.Errorf("account is not a contract account")
 	}
 	return nil
 }