+++ x/erc20/keeper/setup_test.go
@@ -59,10 +59,11 @@ func TestKeeperTestSuite(t *testing.T) {
 	s = new(KeeperTestSuite)
 	suite.Run(t, s)
 
+	// Run Ginkgo integration tests
 	RegisterFailHandler(Fail)
-	RunSpecs(t, "ERC20 Keeper Integration Tests Suite")
+	RunSpecs(t, "Keeper Suite")
 }
 
 func (suite *KeeperTestSuite) SetupTest() {
-	suite.DoSetupTest()
+	suite.DoSetupTest(suite.T())
 }