+++ x/evm/genesis_test.go
@@ -4,14 +4,12 @@ import (
 	"math/big"
 	"testing"
 
+	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/crypto"
-	"github.com/evmos/evmos/v19/contracts"
 	"github.com/evmos/evmos/v19/crypto/ethsecp256k1"
-	testfactory "github.com/evmos/evmos/v19/testutil/integration/evmos/factory"
-	testhandler "github.com/evmos/evmos/v19/testutil/integration/evmos/grpc"
 	testkeyring "github.com/evmos/evmos/v19/testutil/integration/evmos/keyring"
 	testnetwork "github.com/evmos/evmos/v19/testutil/integration/evmos/network"
+	evmostypes "github.com/evmos/evmos/v19/types"
 	"github.com/evmos/evmos/v19/x/evm"
 	"github.com/evmos/evmos/v19/x/evm/statedb"
 	"github.com/evmos/evmos/v19/x/evm/types"
@@ -21,8 +19,6 @@ import (
 type GenesisTestSuite struct {
 	keyring testkeyring.Keyring
 	network *testnetwork.UnitTestNetwork
-	handler testhandler.Handler
-	factory testfactory.TxFactory
 }
 
 func SetupTest() *GenesisTestSuite {
@@ -30,14 +26,9 @@ func SetupTest() *GenesisTestSuite {
 	network := testnetwork.NewUnitTestNetwork(
 		testnetwork.WithPreFundedAccounts(keyring.GetAllAccAddrs()...),
 	)
-	handler := testhandler.NewIntegrationHandler(network)
-	factory := testfactory.New(network, handler)
-
 	return &GenesisTestSuite{
 		keyring: keyring,
 		network: network,
-		handler: handler,
-		factory: factory,
 	}
 }
 
@@ -53,7 +44,6 @@ func TestInitGenesis(t *testing.T) {
 		name     string
 		malleate func(*testnetwork.UnitTestNetwork)
 		genState *types.GenesisState
-		code     common.Hash
 		expPanic bool
 	}{
 		{
@@ -94,6 +84,40 @@ func TestInitGenesis(t *testing.T) {
 			expPanic: true,
 		},
 		{
+			name: "invalid account type",
+			malleate: func(network *testnetwork.UnitTestNetwork) {
+				acc := authtypes.NewBaseAccountWithAddress(address.Bytes())
+				network.App.AccountKeeper.SetAccount(network.GetContext(), acc)
+			},
+			genState: &types.GenesisState{
+				Params: types.DefaultParams(),
+				Accounts: []types.GenesisAccount{
+					{
+						Address: address.String(),
+					},
+				},
+			},
+			expPanic: true,
+		},
+		{
+			name: "invalid code hash",
+			malleate: func(network *testnetwork.UnitTestNetwork) {
+				ctx := network.GetContext()
+				acc := network.App.AccountKeeper.NewAccountWithAddress(ctx, address.Bytes())
+				network.App.AccountKeeper.SetAccount(ctx, acc)
+			},
+			genState: &types.GenesisState{
+				Params: types.DefaultParams(),
+				Accounts: []types.GenesisAccount{
+					{
+						Address: address.String(),
+						Code:    "ffffffff",
+					},
+				},
+			},
+			expPanic: true,
+		},
+		{
 			name: "ignore empty account code checking",
 			malleate: func(network *testnetwork.UnitTestNetwork) {
 				ctx := network.GetContext()
@@ -112,18 +136,21 @@ func TestInitGenesis(t *testing.T) {
 			expPanic: false,
 		},
 		{
-			name: "valid account with code",
+			name: "ignore empty account code checking with non-empty codehash",
 			malleate: func(network *testnetwork.UnitTestNetwork) {
-				ctx := network.GetContext()
-				acc := network.App.AccountKeeper.NewAccountWithAddress(ctx, address.Bytes())
-				network.App.AccountKeeper.SetAccount(ctx, acc)
+				ethAcc := &evmostypes.EthAccount{
+					BaseAccount: authtypes.NewBaseAccount(address.Bytes(), nil, 0, 0),
+					CodeHash:    common.BytesToHash([]byte{1, 2, 3}).Hex(),
+				}
+
+				network.App.AccountKeeper.SetAccount(network.GetContext(), ethAcc)
 			},
 			genState: &types.GenesisState{
 				Params: types.DefaultParams(),
 				Accounts: []types.GenesisAccount{
 					{
 						Address: address.String(),
-						Code:    "1234",
+						Code:    "",
 					},
 				},
 			},
@@ -164,80 +191,7 @@ func TestInitGenesis(t *testing.T) {
 						*tc.genState,
 					)
 				})
-
-				// If the initialization has not panicked we're checking the state
-				for _, account := range tc.genState.Accounts {
-					acc := ts.network.App.AccountKeeper.GetAccount(ctx, common.HexToAddress(account.Address).Bytes())
-					require.NotNil(t, acc, "account not found in account keeper")
-
-					expHash := crypto.Keccak256Hash(common.Hex2Bytes(account.Code))
-					if account.Code == "" {
-						expHash = common.BytesToHash(types.EmptyCodeHash)
-					}
-
-					require.Equal(t,
-						expHash.String(),
-						ts.network.App.EvmKeeper.GetCodeHash(
-							ts.network.GetContext(),
-							common.HexToAddress(account.Address),
-						).String(),
-						"code hash mismatch",
-					)
-
-					require.Equal(t,
-						account.Code,
-						common.Bytes2Hex(
-							ts.network.App.EvmKeeper.GetCode(
-								ts.network.GetContext(),
-								expHash,
-							),
-						),
-						"code mismatch",
-					)
-
-					for _, storage := range account.Storage {
-						key := common.HexToHash(storage.Key)
-						value := common.HexToHash(storage.Value)
-						require.Equal(t, value, vmdb.GetState(common.HexToAddress(account.Address), key), "storage mismatch")
-					}
-				}
 			}
 		})
 	}
-}
-
-func TestExportGenesis(t *testing.T) {
-	ts := SetupTest()
-
-	contractAddr, err := ts.factory.DeployContract(
-		ts.keyring.GetPrivKey(0),
-		types.EvmTxArgs{},
-		testfactory.ContractDeploymentData{
-			Contract:        contracts.ERC20MinterBurnerDecimalsContract,
-			ConstructorArgs: []interface{}{"TestToken", "TTK", uint8(18)},
-		},
-	)
-	require.NoError(t, err, "failed to deploy contract")
-	require.NoError(t, ts.network.NextBlock(), "failed to advance block")
-
-	contractAddr2, err := ts.factory.DeployContract(
-		ts.keyring.GetPrivKey(0),
-		types.EvmTxArgs{},
-		testfactory.ContractDeploymentData{
-			Contract:        contracts.ERC20MinterBurnerDecimalsContract,
-			ConstructorArgs: []interface{}{"AnotherToken", "ATK", uint8(18)},
-		},
-	)
-	require.NoError(t, err, "failed to deploy contract")
-	require.NoError(t, ts.network.NextBlock(), "failed to advance block")
-
-	genState := evm.ExportGenesis(ts.network.GetContext(), ts.network.App.EvmKeeper)
-	require.Len(t, genState.Accounts, 2, "expected only one smart contract in the exported genesis")
-
-	genAddresses := make([]string, 0, len(genState.Accounts))
-	for _, acc := range genState.Accounts {
-		genAddresses = append(genAddresses, acc.Address)
-	}
-	require.Contains(t, genAddresses, contractAddr.Hex(), "expected contract 1 address in exported genesis")
-	require.Contains(t, genAddresses, contractAddr2.Hex(), "expected contract 2 address in exported genesis")
 }