+++ x/evm/statedb/statedb.go
@@ -527,27 +527,18 @@ func (s *StateDB) commitWithCtx(ctx sdk.Context) error {
 		obj := s.stateObjects[addr]
 		if obj.suicided {
 			if err := s.keeper.DeleteAccount(ctx, obj.Address()); err != nil {
-				return errorsmod.Wrapf(err, "failed to delete account %s", obj.Address())
+				return errorsmod.Wrap(err, "failed to delete account")
 			}
 		} else {
 			if obj.code != nil && obj.dirtyCode {
-				if len(obj.code) == 0 {
-					s.keeper.DeleteCode(ctx, obj.CodeHash())
-				} else {
-					s.keeper.SetCode(ctx, obj.CodeHash(), obj.code)
-				}
+				s.keeper.SetCode(ctx, obj.CodeHash(), obj.code)
 			}
 			if err := s.keeper.SetAccount(ctx, obj.Address(), obj.account); err != nil {
 				return errorsmod.Wrap(err, "failed to set account")
 			}
-
 			for _, key := range obj.dirtyStorage.SortedKeys() {
 				valueBytes := obj.dirtyStorage[key].Bytes()
-				if len(valueBytes) == 0 {
-					s.keeper.DeleteState(ctx, obj.Address(), key)
-				} else {
-					s.keeper.SetState(ctx, obj.Address(), key, valueBytes)
-				}
+				s.keeper.SetState(ctx, obj.Address(), key, valueBytes)
 			}
 		}
 	}