+++ x/evm/types/key.go
@@ -28,7 +28,6 @@ const (
 	prefixCode = iota + 1
 	prefixStorage
 	prefixParams
-	prefixCodeHash
 )
 
 // prefix bytes for the EVM transient store
@@ -41,10 +40,9 @@ const (
 
 // KVStore key prefixes
 var (
-	KeyPrefixCode     = []byte{prefixCode}
-	KeyPrefixStorage  = []byte{prefixStorage}
-	KeyPrefixParams   = []byte{prefixParams}
-	KeyPrefixCodeHash = []byte{prefixCodeHash}
+	KeyPrefixCode    = []byte{prefixCode}
+	KeyPrefixStorage = []byte{prefixStorage}
+	KeyPrefixParams  = []byte{prefixParams}
 )
 
 // Transient Store key prefixes